+++ CHANGELOG.rst
@@ -14,7 +14,7 @@ Breaking changes
 ^^^^^^^^^^^^^^^^
 * `ravenpy` now requires `pooch>=1.8.0` for downloading and caching remote testing data. (PR #513)
 * `ravenpy.utilities.testdata` has been refactored to new module `ravenpy.testing`. The `publish_release_notes` function is now located in `ravenpy.utilities.publishing`. (PR #513)
-* The `ravenpy.testing.utils` module now provides a `yangtze()` class for fetching and caching the `raven-testdata` testing data. This replaces the previous `get_local_testdata` and `get_file` functions. (PR #513)
+* The `ravenpy.testing.utils` module now provides a `yangtze()` class for fetching and caching the `raven-testdata` testing data. A convenience function (`get_file`) replaces the previous `get_local_testdata`. (PR #513)
 * The `ravenpy.testing.utils.open_dataset` function no longer supports OPeNDAP URLs or local file paths. Instead, it uses the `yangtze()` class to fetch datasets from the testing data repository or the local cache. Users should now use `xarray.open_dataset()` directly for OPeNDAP URLs or local files. (PR #513)
 
 Bug fixes